// src/FormDesigner/_Common/Enums/WidgetTypeEnum.ts
/**
 * 组件类型枚举
 */
export enum WidgetTypeEnum {
    ROOT = 'root',
    INPUT = 'input',
    TEXTAREA = 'textarea',
    INPUT_NUMBER = 'inputNumber',
    INPUT_PASSWORD = 'inputPassword',
    SELECT = 'select',
    RADIO = 'radio',
    CHECKBOX = 'checkbox',
    DATE_PICKER = 'datePicker',
    TIME_PICKER = 'timePicker',
    DATETIME_PICKER = 'datetimePicker',
    DATE_RANGE_PICKER = 'dateRangePicker',
    TIME_RANGE_PICKER = 'timeRangePicker',
    DATETIME_RANGE_PICKER = 'datetimeRangePicker',
    UPLOAD = 'upload',

    TABLE = 'table',
    LIST = 'list',
    TREE = 'tree',
}